+++ internal/services/tour/delete.go
@@ -4,6 +4,6 @@ import (
 	"context"
 )
 
-func (t *TourService) DeleteTour(ctx context.Context, id int) error {
+func (t *TourService) DeleteTour(ctx context.Context, id int64) error {
 	return t.repo.DeleteTour(ctx, id)
 }
