+++ internal/services/tour/get.go
@@ -5,6 +5,6 @@ import (
 	"localx/internal/models"
 )
 
-func (t *TourService) GetTourById(ctx context.Context, id int) (models.Tour, error) {
+func (t *TourService) GetTourById(ctx context.Context, id int64) (models.Tour, error) {
 	return t.repo.GetTourById(ctx, id)
 }
